Rename `writeback` types to `actions` types
