+++ frontend/src/metabase-types/api/mocks/index.ts
@@ -15,4 +15,4 @@ export * from "./timeline";
 export * from "./settings";
 export * from "./snippets";
 export * from "./user";
-export * from "./writeback";
+export * from "./actions";
